tests/wpt/css-tests/css-exclusions-1_dev/xhtml1/wrap-flow-003.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Exclusions Test: wrap-flow property set to auto</title>
        <link href="mailto:jacobg@adobe.com" rel="author" title="Jacob Goldstein" />
        <link href="http://www.w3.org/TR/css3-exclusions/#wrap-flow-property" rel="help" />
        <meta content="ahem dom" name="flags" />
        <meta content="No exclusion should be created" name="assert" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: auto;
                position: absolute;
                top: 10px;
                left: 20px;
                width: 60px;
                height: 25px;
                background: blue;
            }
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="line1">ethical beer craft</span><br />
            <span id="line2">Fixie pork belly</span><br />
            <span id="line3">food truck in marfa</span><br />
            <span id="line4">american apparel squid</span>
        </div>

        <div id="log"></div>

        <script type="text/javascript">
            test(function() {assert_equals(checkLinePos("line1",216,"right"), true)}, "Verify right of the 'line1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line2",192,"right"), true)}, "Verify right of the 'line2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line3",228,"right"), true)}, "Verify right of the 'line3' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line4",36,"top"), true)}, "Verify top of the 'line4' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line4",264,"right"), true)}, "Verify right of the 'line4' span is positioned correctly");
        </script>
    

</body></html>
